DV-161 feat: 응답 모델 선언 파일 위치를 models에서 schemas로 이동, models는 ai 모델을 선언하는 곳으로 변경
