+++ app/utils/s3.py
@@ -2,7 +2,7 @@ import boto3
 from botocore.exceptions import NoCredentialsError, PartialCredentialsError
 from fastapi import HTTPException
 
-from app.config import Config
+from app.core.config import Config
 
 s3_client = boto3.client(
     "s3",
