+++ pyproject.toml
@@ -22,6 +22,7 @@ boto3 = "^1.35.64"
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.3.3"
 httpx = "^0.27.2"
+tox = "^4.23.2"
 
 [build-system]
 requires = ["poetry-core"]
